+++ .eslintrc
@@ -1,0 +1,12 @@
+root: true
+
+parserOptions:
+  ecmaVersion: 2018
+  sourceType: "module"
+
+env:
+  es2017: true
+  browser: true
+
+extends:
+  "htmlacademy/vanilla"
